CT_ui/requirements.txt
absl-py==2.1.0
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
ament-cmake-test==1.3.6
ament-copyright==0.12.9
ament-cppcheck==0.12.9
ament-cpplint==0.12.9
ament-flake8==0.12.9
ament-index-python==1.4.0
ament-lint==0.12.9
ament-lint-cmake==0.12.9
ament-package==0.14.0
ament-pep257==0.12.9
ament-uncrustify==0.12.9
ament-xmllint==0.12.9
angles==1.15.0
builtin-interfaces==1.2.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
composition-interfaces==1.2.1
contourpy==1.1.1
control-msgs==4.4.0
controller-manager==2.36.0
controller-manager-msgs==2.36.0
cv-bridge==3.2.1
cycler==0.12.1
demo-nodes-py==0.20.3
diagnostic-msgs==4.2.3
diagnostic-updater==3.1.2
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.6
geometry-msgs==4.2.3
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.1
google-auth-oauthlib==1.0.0
grpcio==1.62.0
idna==3.6
image-geometry==3.2.1
importlib-metadata==7.0.1
importlib_resources==6.1.2
interactive-markers==2.3.2
Jinja2==3.1.3
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
kiwisolver==1.4.5
labelImg==1.8.6
laser-geometry==2.4.0
launch==1.0.4
launch-ros==0.19.6
launch-testing==1.0.4
launch-testing-ros==0.19.6
launch-xml==1.0.4
launch-yaml==1.0.4
lifecycle-msgs==1.2.1
logging-demo==0.20.3
lxml==5.1.0
map-msgs==2.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.5
message-filters==4.3.3
mpmath==1.3.0
nav-msgs==4.2.3
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
osrf-pycommon==2.0.2
packaging==23.2
pandas==2.0.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
pillow==10.2.0
prettytable==3.10.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyecharts==2.0.5
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
pytz==2024.1
PyYAML==6.0.1
qt-dotgraph==2.2.2
qt-gui==2.2.2
qt-gui-cpp==2.2.2
qt-gui-py-common==2.2.2
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.11
rcutils==5.1.4
requests==2.31.0
requests-oauthlib==1.3.1
resource-retriever==3.1.1
rmw-dds-common==1.6.0
ros2action==0.18.7
ros2bag==0.15.8
ros2cli==0.18.7
ros2component==0.18.7
ros2controlcli==2.36.0
ros2doctor==0.18.7
ros2interface==0.18.7
ros2launch==0.19.6
ros2lifecycle==0.18.7
ros2multicast==0.18.7
ros2node==0.18.7
ros2param==0.18.7
ros2pkg==0.18.7
ros2run==0.18.7
ros2service==0.18.7
ros2topic==0.18.7
rosbag2-interfaces==0.15.8
rosbag2-py==0.15.8
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt==1.1.5
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.2
rqt-graph==1.3.0
rqt-gui==1.1.5
rqt-gui-py==1.1.5
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.5
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9
scipy==1.10.1
scripts==3.7.0
seaborn==0.13.2
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
shape-msgs==4.2.3
shiboken6==6.6.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sros2==0.10.4
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
sympy==1.12
teleop-twist-keyboard==2.3.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tf-transformations==1.0.0
tf2-geometry-msgs==0.25.5
tf2-kdl==0.25.5
tf2-msgs==0.25.5
tf2-py==0.25.5
tf2-ros-py==0.25.5
tf2-tools==0.25.5
thop==0.1.1.post2209072238
topic-monitor==0.20.3
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
trajectory-msgs==4.2.3
triton==2.2.0
turtlesim==1.4.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.24
unique-identifier-msgs==2.2.1
urllib3==2.2.1
val==0.6
visualization-msgs==4.2.3
wcwidth==0.2.13
Werkzeug==3.0.1
xacro==2.0.8
zipp==3.17.0